Move functions with Gdn dependency to new file
